requirements.txt
pillow>=10.0.0
opencv-python>=4.8.0
pytesseract>=0.3.10
